backend/requirements.txt
# Zynk Backend - Python Dependencies

fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
python-dotenv
openai
requests
python-multipart
faster-whisper